Fixed a bug in filter evaluation when handling an array of shorts. (FELIX-1258)
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@786537 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/src/main/java/org/apache/felix/framework/util/ldap/Parser.java
@@ -1513,11 +1513,11 @@ loop:   for (;;)
         }
         else if (clazz == Short.TYPE)
         {
-            byte[] src = (byte[]) array;
-            array = new Byte[src.length];
+            short[] src = (short[]) array;
+            array = new Short[src.length];
             for (int i = 0; i < src.length; i++)
             {
-                ((Object[]) array)[i] = new Byte(src[i]);
+                ((Object[]) array)[i] = new Short(src[i]);
             }
         }
         else if (clazz == Integer.TYPE)
